make webhook impersonation feature
On the topic of attention economy. Have you ever thought about how damaging that is? Attention is something that we can only spend with Time. They're not taking our money, they're taking our time, not only that but attention is literally the very thing you are thinking of.

If you are paying attention to, well, anything, you will realize that paying attention is very taxing. It's literally what you are doing at any given moment.

It's silly to compare the state of our internet and social media to some comically evil cartoon villain controlling the mind of the masses. But isn't that what's happening here? We consume and forced to think, separated into this echo  chamber where we perpetuate hate and stereotype on every...

Oh, webhook thing? Yeah, I just made it so that the instagram replacer link would replace the original message with a webhook that impersonate the author.