added sending error message to server
But really, what if, what if, I added this to minecraft? Would it work? I mean, it would mean that you're gonna have to have to  run minecraft and AI at the same time, but it's pretty cool that it works, yea?

I wanna ssee if that's possible yknow? Making minecraft mod like this??? I really think that the potential is there! Not to mention we have Touhou Little Maid Mod that makes it possible for anyone to add any waifu of their liking to minecraft.

That'll be hella epic don't you think? I think I should try making that next, who knows???