change the webhook system from auto deleting
Now webhook stays in each channel with a unique name for each character

This means there's a hard limitation of 10 characters per channel

If it ever came to my attention that more than 2 people in this world
uses this discord bot... Then I might just tweak the system again so
that it uses only one webhook per channel

With all that said, have fun then~

Wait, wait, also, I added the delete and edit message function. It only
happpens in the front end tho, in the discord part, but it doesn't show
in the character memory part. I need to redo literally every fucking
thing in terms of character context and memory.

Might even have to use Json to handle memory and system prompting...

...

Well fuck me, this joke project turned out to be one of my biggest one
yet!